+++ tests/acceptance/pageObjects/FilesPageElement/publicLinksCreationForm.js
@@ -1,0 +1,25 @@
+const util = require('util')
+
+module.exports = {
+  commands: {
+    /**
+     * sets up the xpath for year in expiry date of public link
+     *
+     * @param year
+     * @returns {{locateStrategy: string, selector: *}}
+     */
+    setExpiryDateYearSelectorXpath: function (year) {
+      const yearSelectorXpath = util.format(this.elements.dateTimeYearPicker.selector, year)
+      return {
+        selector: yearSelectorXpath,
+        locateStrategy: this.elements.dateTimeYearPicker.locateStrategy
+      }
+    }
+  },
+  elements: {
+    dateTimeYearPicker: {
+      selector: '//div[@class="vdatetime-year-picker"]//div[normalize-space(.)="%s"]',
+      locateStrategy: 'xpath'
+    }
+  }
+}
